added template clients
